Available algorithms

Candidate generation based

TreeProjection based

Available algorithms

Candidate generation based

Missing docstring.

Missing docstring for apriori(miner::Miner, X::AbstractDataset; verbose::Bool=false). Check Documenter's build log for details.

TreeProjection based

FPGrowth

Missing docstring.

Missing docstring for fpgrowth(miner::Miner, X::AbstractDataset; verbose::Bool=false). Check Documenter's build log for details.

FPGrowth algorithm relies on two data structures, FPTree and HeaderTable. To know more about them and their, please refer to the documentation here data-structures.

FPGrowth algorithm relies on the following two routines.

ModalAssociationRules.patternbase — Method

patternbase(
    item::Item,
    htable::HeaderTable,
    miner::AbstractMiner
)::ConditionalPatternBase

Retrieve the ConditionalPatternBase of fptree based on item.

The conditional pattern based on a FPTree is the set of all the paths from the tree root to nodes containing item (not included). Each of these paths is represented by an EnhancedItemset.

The EnhancedItemsets in the returned ConditionalPatternBase are sorted decreasingly by gsupport.

See also AbstractMiner, ConditionalPatternBase, contributors, EnhancedItemset, fpgrowth, FPTree, Item, Itemset, WorldMask.

ModalAssociationRules.projection — Method

function projection(pbase::ConditionalPatternBase, miner::AbstractMiner)

Return respectively a FPTree and a HeaderTable starting from pbase. An AbstractMiner must be provided to guarantee the generated header table internal state is OK, that is, its items are sorted decreasingly by gsupport.

See also ConditionalPatternBase, FPTree, gsupport, HeaderTable, AbstractMiner.

Also, FPGrowth requires the Miner to remember the Contributors associated with the extracted frequent itemsets. To add this functionality, we can define a new dispatch of initminingstate: it is automatically considered to enrich the miner, while building it together with fpgrowth as mining algorithm.

ModalAssociationRules.initminingstate — Method

initminingstate(::typeof(fpgrowth), ::MineableData)::MiningState

MiningState fields levereged when executing FP-Growth algorithm.

See also hasminingstate, MiningState, miningstate.