Built in data structures

FPTree

HeaderTable

Built in data structures

FPTree

SoleRules.FPTree — Type

mutable struct FPTree
    content::Union{Nothing,Item}        # Item contained in this node (nothing if root)

    parent::Union{Nothing,FPTree}       # parent node
    const children::Vector{FPTree}      # children nodes

    count::Int64                        # number of equal Items this node represents

    # how many times lsupp(content) does overpass
    # the corresponding threshold for each world
    const contributors::WorldMask

    link::Union{Nothing,FPTree}         # link to another FPTree root
end

Fundamental data structure used in FP-Growth algorithm. Essentialy, an FPTree is a prefix tree where a root-leaf path represent an Itemset.

Consider the Itemsets sorted by gsupport of their items. An FPTree is such that the common Items-prefix shared by different Itemsets is not stored multiple times.

This implementation generalizes the propositional logic case scenario to modal logic; given two Itemsets sharing a Item prefix, the worlds in which they are true is accumulated.

Info

Did you notice? One FPTree structure contains all the information needed to construct an EnhancedItemset. This is crucial to generate new FPTree during fpgrowth algorithm, via building ConditionalPatternBase iteratively while avoiding visiting the dataset over and over again.

See also EnhancedItemset, fpgrowth, gsupport, Item, Itemset, WorldMask.

SoleRules.content — Method

content(fptree::FPTree)::Union{Nothing,Item}

Getter for the Item (possibly empty) wrapped by fptree.

See also content!, FPTree.

SoleRules.content! — Method

content!(fptree::FPTree, item::Union{Nothing,Item})

Setter for fptree's content (the wrapped item).

See also content, FPTree.

SoleRules.parent — Method

parent(fptree::FPTree)::Union{Nothing,FPTree}

Getter for the parent FPTrees of fptree.

See also FPTree, parent!.

SoleRules.parent! — Method

parent!(fptree::FPTree, item::Union{Nothing,FPTree})

Setter for fptree's parent FPTree.

See also FPTree, parent.

SoleRules.children — Method

children(fptree::FPTree)::Vector{FPTree}

Getter for the list of children FPTrees of fptree.

See also children!, FPTree.

SoleRules.children! — Method

children!(fptree::FPTree, child::FPTree)

Add a new FPTree to fptree's children vector.

Warning

This method forces the new children to be added: it is a caller's responsability to check whether child is not already a children of fptree and, if so, handle the case. This check is performed, for example, in Base.push!.

Note

This method already sets the new children parent to fptree itself.

See also Base.push!, children, FPTree.

Base.count — Method

Base.count(fptree::FPTree)::Int64

Getter for the fptree internal counter. Essentially, it represents the number of overlappings Item which ended up in fptree node during the building process of the tree itself.

See also count!, FPTree, Item.

SoleRules.count! — Method

count!(fptree::FPTree, newcount::Int64)

Setter for fptree's internal counter to a fixed value newcount.

See also count, FPTree.

SoleRules.addcount! — Method

addcount!(fptree::FPTree, newcount::Int64)

Add newcount to fptree's internal counter.

See also count, FPTree.

SoleRules.contributors — Method

contributors(fptree::FPTree)::WorldMask

Getter for the fptree contributors array.

Consider the Contributors definition. In the specific case of an FPTree, the contributors array is simply a vector of integers which answers the following question for each i-th world of a generic instance: given a local support threshold t, how many times is lsupp(content) >= t ?

Essentially, it represents the number of overlappings Item which ended up in fptree node during the building process of the tree itself.

See also Contributors, contributors!, FPTree, Item, lsupport.

SoleRules.contributors! — Method

contributors!(fptree::FPTree, contribution::WorldMask)

Setter for fptree's internal contributors mask to contribution WorldMask.

See also contributors, FPTree, WorldMask.

SoleRules.addcontributors! — Method

addcontributors!(fptree::FPTree, contribution::WorldMask) =

Add the contribution WorldMask to fptree's internal contributors mask.

See also contributors, FPTree, WorldMask.

SoleRules.link — Method

link(fptree::FPTree)::Union{Nothing,FPTree}

Getter for fptree's next brother FPTree. fptree's brotherhood is the set of all the FPTree whose content is exactly fptree.content.

See also content, FPTree.

SoleRules.link! — Method

function link!(from::FPTree, to::FPTree)

Establish a link between two FPTrees. If the starting tree is already linked with something, the already existing link are followed until a new "empty-linked" FPTree is found.

See also follow, FPTree, HeaderTable.

SoleRules.follow — Method

function follow(fptree::FPTree)::Union{Nothing,FPTree}

Follow fptree link to (an internal node of) another FPTree.

See also FPTree, HeaderTable.

SoleRules.islist — Method

islist(fptree::FPTree)::Bool

Return true if every subtree in fptree has exactly 0 or 1 children.

See also FPTree

SoleRules.retrieveall — Method

function retrieveall(fptree::FPTree)::Itemset

Return all the unique Items appearing in fptree.

See also FPTree, Item, Itemset.

Base.push! — Method

function Base.push!(
    fptree::FPTree,
    itemset::Itemset,
    ninstance::Int64,
    miner::Miner;
    htable::Union{Nothing,HeaderTable}=nothing
)

function Base.push!(
    fptree::FPTree,
    itemset::EnhancedItemset,
    ninstance::Int64,
    miner::Miner;
    htable::Union{Nothing,HeaderTable}=nothing
)

Base.push!(
    fptree::FPTree,
    enhanceditemsets::ConditionalPatternBase,
    miner::Miner;
    htable::Union{Nothing,HeaderTable}=nothing
)

Push one or more Itemsets/EnhancedItemset to an FPTree. If an HeaderTable is provided, it is leveraged to develop internal links.

Warning

To optimally leverage the compression capabilities of FPTrees, the Itemsets provided should be sorted decreasingly by gsupport. By default, to improve performances, this check is not performed inside this method.

See also EnhancedItemset, FPTree, gsupport, HeaderTable, Itemset.

HeaderTable

SoleRules.HeaderTable — Type

struct HeaderTable
    items::Vector{Item}
    link::Dict{Item,Union{Nothing,FPTree}}
end

Utility data structure used to fastly access FPTree internal nodes.

SoleRules.items — Method

items(htable::HeaderTable)::Vector{Item}

Getter for the Items loaded inside htable.

See also HeaderTable, Item.

SoleRules.link — Method

link(htable::HeaderTable)
link(htable::HeaderTable, item::Item)

Getter for the link structure wrapped by htable, or one of its specific entry.

The link structure is, essentially, a dictionary associating an Item to a specific FPTree.

See also FPTree, HeaderTable, Item, link!.

SoleRules.follow — Method

function follow(htable::HeaderTable, item::Item)::Union{Nothing,FPTree}

Follow htable link to (an internal node of) a FPTree.

See also FPTree, HeaderTable, Item, link, link!.

SoleRules.link! — Method

function link!(htable::HeaderTable, fptree::FPTree)

Establish a link towards fptree, following the entry in htable corresponding to the content of fptree.

See also content, FPTree, HeaderTable.

SoleRules.checksanity! — Method

function checksanity!(htable::HeaderTable, miner::Miner)::Bool

Check if htable internal state is correct, that is, its items are sorted decreasingly by global support. If items are already sorted, return true; otherwise, sort them and return false.

See also Miner, gsupport, HeaderTable, items.

Base.reverse — Method

Base.reverse(htable::HeaderTable)

Iterator on htable wrapped Items, in reverse order.

See also HeaderTable, Item.