Getting started

Core definitions

Measures

Mining structures

Getting started

In this introductory section you will learn about the main building blocks of SoleRules.

Core definitions

SoleRules.Item — Type

const Item = SoleLogics.Formula

Fundamental type in the context of association rule mining. An Item is a logical formula, which can be SoleLogics.checked on models.

The purpose of association rule mining (ARM) is to discover interesting relations between Items, regrouped in Itemsets, to generate association rules (ARule).

Interestingness is established through a set of MeaningfulnessMeasure.

See also SoleLogics.check, gconfidence, lsupport, MeaningfulnessMeasure, SoleLogics.Formula.

SoleRules.Itemset — Type

const Itemset = Vector{Item}

Collection of unique Items.

Given a MeaningfulnessMeasure meas and a threshold to be overpassed, t, then an itemset itemset is said to be meaningful with respect to meas if and only if meas(itemset) > t. Alternatively to meaningful, it is said to be frequent.

Generally speaking, meaningfulness (or interestingness) of an itemset is directly correlated to its frequency in the data: intuitively, when a pattern is recurrent in data, then it is candidate to be interesting.

Every association rule mining algorithm aims to find frequent itemsets by applying meaningfulness measures such as local and global support, respectively lsupport and gsupport.

Frequent itemsets are then used to generate association rules (ARule).

See also ARule, gsupport, Item, lsupport, MeaningfulnessMeasure.

SoleRules.toformula — Function

toformula(itemset::Itemset)

Conjunctive normal form of the Items contained in itemset.

See also Item, Itemset, SoleLogics.LeftmostConjunctiveForm

SoleRules.Threshold — Type

const Threshold = Float64

Threshold value for meaningfulness measures.

See also gconfidence, gsupport, lconfidence, lsupport.

SoleRules.ARule — Type

const ARule = Tuple{Itemset,Itemset}

An association rule represents a strong and meaningful co-occurrence relationship between two Itemsets whose intersection is empty.

Generating all the ARule "hidden" in the data is the main purpose of ARM.

The general framework always followed by ARM techniques is to, firstly, generate all the frequent itemsets considering a set of MeaningfulnessMeasure specifically tailored to work with Itemsets. Thereafter, all the association rules are generated by testing all the combinations of frequent itemsets against another set of MeaningfulnessMeasure, this time designed to capture how "reliable" a rule is.

See also gconfidence, Itemset, lconfidence, MeaningfulnessMeasure.

SoleRules.content — Method

content(rule::ARule)::Tuple{Itemset,Itemset}

Getter for the content of an ARule, that is, both its antecedent and its consequent.

See also antecedent, ARule, consequent, Itemset.

SoleRules.antecedent — Method

antecedent(rule::ARule)::Itemset

Getter for rule's antecedent.

See also antecedent, ARule, Itemset.

SoleRules.consequent — Method

consequent(rule::ARule)::Itemset

Getter for rule's consequent.

See also consequent, ARule, Itemset.

SoleRules.ARMSubject — Type

ARMSubject = Union{ARule,Itemset}

Memoizable types for association rule mining (ARM).

See also GmeasMemo, GmeasMemoKey, LmeasMemo, LmeasMemoKey.

Measures

SoleRules.MeaningfulnessMeasure — Type

const MeaningfulnessMeasure = Tuple{Function, Threshold, Threshold}

In the classic propositional case scenario where each instance of a dataset is composed of just a single world (it is a propositional interpretation), a meaningfulness measure is simply a function which measures how many times a property of an Itemset or an ARule is respected across all instances of the dataset.

In the context of modal logic, where the instances of a dataset are relational objects, every meaningfulness measure must capture two aspects: how much an Itemset or an ARule is meaningful inside an instance, and how much the same object is meaningful across all the instances.

For this reason, we can think of a meaningfulness measure as a matryoshka composed of an external global measure and an internal local measure. The global measure tests for how many instances a local measure overpass a local threshold. At the end of the process, a global threshold can be used to establish if the global measure is actually meaningful or not. (Note that this generalizes the propositional logic case scenario, where it is enough to just apply a measure across instances.)

Therefore, a MeaningfulnessMeasure is a tuple composed of a global meaningfulness measure, a local threshold and a global threshold.

See also gconfidence, gsupport, lconfidence, lsupport.

SoleRules.islocalof — Method

islocalof(::Function, ::Function)::Bool

Twin method of isglobalof.

Trait to indicate that a local meaningfulness measure is used as subroutine in a global measure.

For example, islocalof(lsupport, gsupport) is true, and isglobalof(gsupport, lsupport) is false.

Warning

When implementing a custom meaningfulness measure, make sure to implement both traits if necessary. This is fundamental to guarantee the correct behavior of some methods, such as getlocalthreshold.

See also getlocalthreshold, gsupport, isglobalof, lsupport.

SoleRules.isglobalof — Method

isglobalof(::Function, ::Function)::Bool

Twin trait of islocalof.

See also getlocalthreshold, gsupport, islocalof, lsupport.

SoleRules.LmeasMemoKey — Type

const LmeasMemoKey = Tuple{Symbol,ARMSubject,Int64}

Key of a LmeasMemo dictionary. Represents a local meaningfulness measure name (as a Symbol), a ARMSubject, and the number of a dataset instance where the measure is applied.

See also LmeasMemo, ARMSubject.

SoleRules.LmeasMemo — Type

const LmeasMemo = Dict{LmeasMemoKey,Threshold}

Association between a local measure of a ARMSubject on a specific dataset instance, and its value.

See also LmeasMemoKey, ARMSubject.

SoleRules.GmeasMemoKey — Type

const GmeasMemoKey = Tuple{Symbol,ARMSubject}

Key of a GmeasMemo dictionary. Represents a global meaningfulness measure name (as a Symbol) and a ARMSubject.

See also GmeasMemo, ARMSubject.

SoleRules.GmeasMemo — Type

const GmeasMemo = Dict{GmeasMemoKey,Threshold}

Association between a global measure of a ARMSubject on a dataset, and its value.

The reference to the dataset is not explicited here, since GmeasMemo is intended to be used as a memoization structure inside ARuleMiner objects, and the latter already knows the dataset they are working with.

See also GmeasMemoKey, ARMSubject.

What follows is a list of the already built-in meaningfulness measures. In the hands-on section you will learn how to implement your own measure.

SoleRules.lsupport — Method

function lsupport(
    itemset::Itemset,
    logi_instance::LogicalInstance;
    miner::Union{Nothing,ARuleMiner}=nothing
)::Float64

Compute the local support for the given itemset in the given logi_instance.

Local support is the ratio between the number of worlds in a LogicalInstance where and Itemset is true and the total number of worlds in the same instance.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also ARuleMiner, LogicalInstance, Itemset.

SoleRules.gsupport — Method

function gsupport(
    itemset::Itemset,
    X::SupportedLogiset,
    threshold::Threshold;
    miner::Union{Nothing,ARuleMiner} = nothing
)::Float64

Compute the global support for the given itemset on a logiset X, considering threshold as the threshold for the local support called internally.

Global support is the ratio between the number of LogicalInstances in a SupportedLogiset for which the local support, lsupport, is greater than a Threshold, and the total number of instances in the same logiset.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also ARuleMiner, LogicalInstance, Itemset, SupportedLogiset, Threshold.

SoleRules.lconfidence — Method

function lconfidence(
    rule::ARule,
    logi_instance::LogicalInstance;
    miner::Union{Nothing,ARuleMiner} = nothing
)::Float64

Compute the local confidence for the given rule in the given logi_instance.

Local confidence is the ratio between lsupport of an ARule on a LogicalInstance and the lsupport of the antecedent of the same rule.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also antecedent, ARule, ARuleMiner, LogicalInstance, lsupport.

SoleRules.gconfidence — Method

function gconfidence(
    rule::ARule,
    X::SupportedLogiset,
    threshold::Threshold;
    miner::Union{Nothing,ARuleMiner} = nothing
)::Float64

Compute the global confidence for the given rule on a logiset X, considering threshold as the threshold for the global support called internally.

Global confidence is the ratio between gsupport of an ARule on a SupportedLogiset and the gsupport of the antecedent of the same rule.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also antecedent, ARule, ARuleMiner, gsupport, SupportedLogiset.

Mining structures

SoleRules.ARuleMiner — Type

struct ARuleMiner
    X::AbstractDataset              # target dataset
    algo::MiningAlgo                # algorithm used to perform extraction

    items::Vector{Item}

    # global meaningfulness measures and their thresholds (both local and global)
    item_constrained_measures::Vector{<:MeaningfulnessMeasure}
    rule_constrained_measures::Vector{<:MeaningfulnessMeasure}

    freqitems::Vector{Itemset}      # collected frequent itemsets
    arules::Vector{ARule}           # collected association rules

    lmemo::LmeasMemo                # local memoization structure
    gmemo::GmeasMemo                # global memoization structure
    info::NamedTuple                # general informations
end

Machine learning model interface to perform association rules extraction.

Examples

julia> using SoleRules
julia> using SoleData

# Load NATOPS DataFrame
julia> X_df, y = load_arff_dataset("NATOPS");

# Convert NATOPS DataFrame to a Logiset
julia> X = scalarlogiset(X_df)

# Prepare some propositional atoms
julia> p = Atom(ScalarCondition(UnivariateMin(1), >, -0.5))
julia> q = Atom(ScalarCondition(UnivariateMin(2), <=, -2.2))
julia> r = Atom(ScalarCondition(UnivariateMin(3), >, -3.6))

# Prepare modal atoms using later relationship - see [`SoleLogics.IntervalRelation`](@ref))
julia> lp = box(IA_L)(p)
julia> lq = diamond(IA_L)(q)
julia> lr = boxlater(r)

# Compose a vector of items, regrouping the atoms defined before
julia> manual_alphabet = Vector{Item}([p, q, r, lp, lq, lr])

# Create an association rule miner wrapping `apriori` algorithm - see [`apriori`](@ref);
# note that meaningfulness measures are not explicited and, thus, are defaulted as in the
# call below.
julia> miner = ARuleMiner(X, apriori(), manual_alphabet)

# Create an association rule miner, expliciting global meaningfulness measures with their
# local and global thresholds, both for [`Itemset`](@ref)s and [`ARule`](@ref).
julia> miner = ARuleMiner(X, apriori(), manual_alphabet,
    [(gsupport, 0.1, 0.1)], [(gconfidence, 0.2, 0.2)])

# Consider the dataset and learning algorithm wrapped by `miner` (resp., `X` and `apriori`)
# Mine the frequent itemsets, that is, those for which item measures are large enough.
# Then iterate the generator returned by [`mine`](@ref) to enumerate association rules.
julia> for arule in SoleRules.mine(miner)
    println(miner)
end

See also ARule, apriori, MeaningfulnessMeasure, Itemset, GmeasMemo, LmeasMemo, MiningAlgo.

SoleRules.MiningAlgo — Type

const MiningAlgo = FunctionWrapper{Nothing,Tuple{ARuleMiner,AbstractDataset}}

Function wrapper representing a function which performs mining of frequent Itemsets in a ARuleMiner.

See also SoleLogics.AbstractDataset, ARuleMiner, apriori, fpgrowth.

SoleRules.dataset — Method

dataset(miner::ARuleMiner)::AbstractDataset

Getter for the dataset wrapped by miners.

See SoleBase.AbstractDataset, ARuleMiner.

SoleRules.algorithm — Method

algorithm(miner::ARuleMiner)::MiningAlgo

Getter for the mining algorithm loaded into miner.

See ARuleMiner, MiningAlgo.

SoleRules.items — Method

items(miner::ARuleMiner)

Getter for the items of Items loaded into miner.

See ARuleMiner, Item, MiningAlgo.

SoleRules.item_meas — Method

item_meas(miner::ARuleMiner)::Vector{<:MeaningfulnessMeasure}

Return the MeaningfulnessMeasures tailored to work with Itemsets, loaded inside miner.

See ARuleMiner, Itemset, MeaningfulnessMeasure

SoleRules.rule_meas — Method

rule_meas(miner::ARuleMiner)::Vector{<:MeaningfulnessMeasure}

Return the MeaningfulnessMeasures tailored to work with ARules, loaded inside miner.

See ARuleMiner, ARule, MeaningfulnessMeasure.

SoleRules.getlocalthreshold — Method

getlocalthreshold(miner::ARuleMiner, meas::Function)::Threshold

Getter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work locally (that is, analyzing "the inside" of a dataset's instances) in miner.

See ARuleMiner, MeaningfulnessMeasure, Threshold.

SoleRules.setlocalthreshold — Method

setlocalthreshold(miner::ARuleMiner, meas::Function, threshold::Threshold)

Setter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work locally (that is, analyzing "the inside" of a dataset's instances) in miner.

See ARuleMiner, MeaningfulnessMeasure, Threshold.

SoleRules.getglobalthreshold — Method

getglobalthreshold(miner::ARuleMiner, meas::Function)::Threshold

Getter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work globally (that is, measuring the behavior of a specific local-measure across all dataset's instances) in miner.

See ARuleMiner, MeaningfulnessMeasure, Threshold.

SoleRules.setglobalthreshold — Method

setglobalthreshold(miner::ARuleMiner, meas::Function, threshold::Threshold)

Setter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work globally (that is, measuring the behavior of a specific local-measure across all dataset's instances) in miner.

See ARuleMiner, MeaningfulnessMeasure, Threshold.

SoleRules.freqitems — Method

freqitems(miner::ARuleMiner)

Return all frequent Itemsets mined by miner.

See also ARuleMiner, Itemset.

SoleRules.arules — Method

arules(miner::ARuleMiner)

Return all the ARules mined by miner.

See also ARule, ARuleMiner.

SoleRules.localmemo — Method

localmemo(miner::ARuleMiner)::LmeasMemo
localmemo(miner::ARuleMiner, key::LmeasMemoKey)

Return the local memoization structure inside miner, or a specific entry if a LmeasMemoKey is provided.

See also ARuleMiner, LmeasMemo, LmeasMemoKey.

SoleRules.localmemo! — Method

localmemo!(miner::ARuleMiner, key::LmeasMemoKey, val::Threshold)

Setter for a specific entry key inside the local memoization structure wrapped by miner.

See also ARuleMiner, LmeasMemo, LmeasMemoKey.

SoleRules.globalmemo — Method

globalmemo(miner::ARuleMiner)::GmeasMemo
globalmemo(miner::ARuleMiner, key::GmeasMemoKey)

Return the global memoization structure inside miner, or a specific entry if a GmeasMemoKey is provided.

See also ARuleMiner, GmeasMemo, GmeasMemoKey.

SoleRules.globalmemo! — Method

globalmemo!(miner::ARuleMiner, key::GmeasMemoKey, val::Threshold)

Setter for a specific entry key inside the global memoization structure wrapped by miner.

See also ARuleMiner, GmeasMemo, GmeasMemoKey.

SoleRules.info — Method

info(miner::ARuleMiner)::NamedTuple
info(miner::ARuleMiner, key::Symbol)

Getter for the entire additional informations field inside a miner, or one of its specific entries.

See also ARuleMiner, NamedTuple.

SoleRules.mine — Method

mine(miner::ARuleMiner)

Synonym for SoleRules.apply(miner, dataset(miner)).

See also ARule, Itemset, SoleRules.apply.

SoleRules.apply — Method

apply(miner::ARuleMiner, X::AbstractDataset)

Extract association rules in the dataset referenced by miner, saving the interesting Itemsets inside miner. Then, return a generator of ARuless.

See also ARule, Itemset.