Association rule mining with modal logic

New building blocks

Modal logic in action

Association rule mining with modal logic

New building blocks

SoleRules.WorldMask — Type

const WorldMask = Vector{Int64}

Vector whose i-th position stores how many times a certain MeaningfulnessMeasure applied on a specific Itemsets is true on the i-th world of multiple instances.

If a single instance is considered, then this acts as a bit mask.

For example, if we consider 5 instances, each of which containing 3 worlds, then the worlds mask of an itemset could be [5,2,0], meaning that the itemset is always true on the first world of every instance. If we consider the second world, the same itemset is true on it only on two instances. If we consider the third world, then the itemset is never true.

See also Contributors, Itemset, MeaningfulnessMeasure.

SoleRules.EnhancedItemset — Type

const EnhancedItemset = Vector{Tuple{Item,Int64,WorldMask}}

"Enhanced" representation of an Itemset, in which each Item is associated to a counter and a specific WorldMask.

Consider an Item called item. The first counter keeps the value of gsupport applied on item itself. The second counter counts on which worlds item is true.

Intuitively, this type is useful to represent and manipulate collections of items when we want to avoid iterating an entire dataset multiple times when extracting frequent Itemset.

Info

To give you a better insight into where this type of data is used, this is widely used behind the scenes in the implementation of fpgrowth, which is the state of art algorithm to perform ARM.

See also fpgrowth, Item, Itemset, WorldMask.

SoleRules.ConditionalPatternBase — Type

const ConditionalPatternBase = Vector{EnhancedItemset}

Collection of EnhancedItemset. This is useful to manipulate certain data structures when looking for frequent Itemsets, such as FPTree.

This is used to implement fpgrowth algorithm as described here.

See also EnhancedItemset, fpgrowth, FPTree.

Modal logic in action

SoleRules.Contributors — Type

Structure for storing association between a local measure, applied on a certain ARMSubject on a certain LogicalInstance, and a vector of integers representing the worlds for which the measure is greater than a certain threshold.

This type is intended to be used inside a Miner info named tuple, to support the execution of, for example, fpgrowth algorthm.

See also LmeasMemoKey, WorldMask

Missing docstring.

Missing docstring for @equip_contributors. Check Documenter's build log for details.

Missing docstring.

Missing docstring for isequipped. Check Documenter's build log for details.

SoleRules.contributors — Function

contributors(fptree::FPTree)::WorldMask

Getter for the fptree contributors array.

Consider the Contributors definition. In the specific case of an FPTree, the contributors array is simply a vector of integers which answers the following question for each i-th world of a generic instance: given a local support threshold t, how many times is lsupp(content) >= t ?

Essentially, it represents the number of overlappings Item which ended up in fptree node during the building process of the tree itself.

See also Contributors, contributors!, FPTree, Item, lsupport.

SoleRules.contributors! — Function

contributors!(miner::Miner, key::LmeasMemoKey, mask::WorldMask)

Set a miner's contributors entry.

See also Miner, LmeasMemoKey, @equip_contributors, WorldMask.

contributors!(fptree::FPTree, contribution::WorldMask)

Setter for fptree's internal contributors mask to contribution WorldMask.

See also contributors, FPTree, WorldMask.