Getting started

Core definitions

Measures

Mining structures

Utilities

The following utilities often involve performing some combinatoric trick between Itemsets and ARules, and might be useful to avoid reinventing the wheel.

Items and Itemsets

SoleRules.combine — Function

combine(itemsets::Vector{<:Itemset}, newlength::Integer)

Return a generator which combines Itemsets from itemsets into new itemsets of length newlength by taking all combinations of two itemsets and joining them.

See also Itemset.

combine(variable::Vector{<:Item}, fixed::Vector{<:Item})

Return a generator of Itemset, which iterates the combinations of Items in variable and prepend them to fixed vector.

See also Item, Itemset.

SoleRules.grow_prune — Function

grow_prune(candidates::Vector{Itemset}, frequents::Vector{Itemset}, k::Integer)

Return a generator, which yields only the candidates for which every (k-1)-length subset is in frequents.

Warning

Generated Itemsets could contain repetitions.

See also Itemset.

SoleRules.coalesce_contributors — Method

coalesce_contributors(
    itemset::Itemset,
    miner::ARuleMiner;
    lmeas::Function=lsupport
)

Consider all the contributors of an ARMSubject on all the instances. Return their sum and a boolean value, indicating whether the resulting contributors overpasses the local support threshold enough times.

See also ARMSubject, contributors, Threshold.

Association rules

SoleRules.arules_generator — Method

arules_generator(itemsets::Vector{Itemset}, miner::ARuleMiner)

Generates association rules from the given collection of itemsets and miner. Iterates through the powerset of each itemset to generate meaningful ARule.

To establish the meaningfulness of each association rule, check if it meets the global constraints specified in rule_meas(miner), and yields the rule if so.

See also ARule, ARuleMiner, Itemset, rule_meas.