Available algorithms

Candidate generation based

TreeProjection based

Available algorithms

Candidate generation based

Missing docstring.

Missing docstring for apriori(;fulldump::Bool=true, verbose::Bool=true). Check Documenter's build log for details.

TreeProjection based

FPGrowth

Missing docstring.

Missing docstring for fpgrowth(;fulldump::Bool=true, verbose::Bool=true). Check Documenter's build log for details.

FPGrowth algorithm relies on two data structures, FPTree and HeaderTable. To know more about them and their, please refer to the documentation here data-structures.

FPGrowth algorithm relies on the following two routines.

SoleRules.patternbase — Method

patternbase(item::Item, htable::HeaderTable, miner::Miner)::ConditionalPatternBase

Retrieve the ConditionalPatternBase of fptree based on item.

The conditional pattern based on a FPTree is the set of all the paths from the tree root to nodes containing item (not included). Each of these paths is represented by an EnhancedItemset, where each Item is associated with a WorldMask, given by the minimum of its contributors and the ones of item.

The EnhancedItemsets in the returned ConditionalPatternBase are sorted decreasingly by gsupport.

See also Miner, ConditionalPatternBase, contributors, EnhancedItemset, fpgrowth, FPTree, Item, Itemset, WorldMask.

SoleRules.projection — Method

function projection(
    pbase::ConditionalPatternBase;
    miner::Union{Nothing,Miner}=nothing
)

Return respectively a FPTree and a HeaderTable starting from pbase. It is reccomended to also provide a Miner to guarantee the generated header table internal state is OK, that is, its items are sorted decreasingly by gsupport.

See also ConditionalPatternBase, FPTree, gsupport, HeaderTable.

Also, FPGrowth requires the Miner to remember the Contributors associated with the extracted frequent itemsets. To add this functionality, we can define a new dispatch of initpowerups: it is automatically considered to enrich the miner, while building it together with fpgrowth as mining algorithm.

SoleRules.initpowerups — Method

initpowerups(::typeof(fpgrowth), ::AbstractDataset)::NamedTuple

Powerups suite for FP-Growth algorithm.

When initializing a Miner with fpgrowth algorithm, this defines how miner's powerup field is filled to optimize the mining.

See also haspowerup, powerup.