Getting started

Core definitions

Measures

Mining structures

Getting started

In this introductory section you will learn about the main building blocks of SoleRules.jl. Also if a good picture about association rule mining (ARM, from now onwards) is given during the documentation, to make the most out of this documentation we suggest to read the following articles:

association rule mining introduction and Apriori algorithm

FPGrowth algorithm

Those up above introduce two important algorithms, which are also built-in in this package. Moreover, the latter one is the state-of-the-art in the field of ARM.

Further on in the documentation, the potential of SoleRules.jl will emerge: this package's raison d'être is to generalize the already existing ARM algorithms to modal logics, which are more expressive than propositional one and computationally less expensive than first order logic. If you are new to Sole.jl and you want to learn more about modal logic, please have a look at SoleLogics.jl for a general overview on the topic, or follow this documentation and return to this link if needed.

Core definitions

One Item is just a logical formula, which can be interpreted by a certain model. At the moment, here, we don't care about how models are represented by Sole.jl under the hood, or how interpretation algorithm works: what matters is that Items are manipulated by ARM algorithms, which try to find which conjunctions between items are most statistically significant.

SoleRules.Item — Type

const Item = SoleLogics.Formula

Fundamental type in the context of association rule mining. An Item is a logical formula, which can be SoleLogics.checked on models.

The purpose of association rule mining (ARM) is to discover interesting relations between Items, regrouped in Itemsets, to generate association rules (ARule).

Interestingness is established through a set of MeaningfulnessMeasure.

See also SoleLogics.check, gconfidence, lsupport, MeaningfulnessMeasure, SoleLogics.Formula.

SoleRules.Itemset — Type

const Itemset = Vector{Item}

Collection of unique Items.

Given a MeaningfulnessMeasure meas and a threshold to be overpassed, t, then an itemset itemset is said to be meaningful with respect to meas if and only if meas(itemset) > t. Alternatively to meaningful, it is said to be frequent.

Generally speaking, meaningfulness (or interestingness) of an itemset is directly correlated to its frequency in the data: intuitively, when a pattern is recurrent in data, then it is candidate to be interesting.

Every association rule mining algorithm aims to find frequent itemsets by applying meaningfulness measures such as local and global support, respectively lsupport and gsupport.

Frequent itemsets are then used to generate association rules (ARule).

See also ARule, gsupport, Item, lsupport, MeaningfulnessMeasure.

Notice that one Itemset could be a set, but actually it is a vector: this is because, often, ARM algorithms need to establish an order between items in itemsets to work efficiently. To convert an Itemset in its conjunctive normla form we simply call toformula.

SoleRules.toformula — Function

toformula(itemset::Itemset)

Conjunctive normal form of the Items contained in itemset.

See also Item, Itemset, SoleLogics.LeftmostConjunctiveForm

Enough about Itemsets. Our final goal is to produce association rules.

SoleRules.ARule — Type

const ARule = Tuple{Itemset,Itemset}

An association rule represents a strong and meaningful co-occurrence relationship between two Itemsets whose intersection is empty.

Generating all the ARule "hidden" in the data is the main purpose of ARM.

The general framework always followed by ARM techniques is to, firstly, generate all the frequent itemsets considering a set of MeaningfulnessMeasure specifically tailored to work with Itemsets. Thereafter, all the association rules are generated by testing all the combinations of frequent itemsets against another set of MeaningfulnessMeasure, this time designed to capture how "reliable" a rule is.

See also gconfidence, Itemset, lconfidence, MeaningfulnessMeasure.

SoleRules.content — Method

content(rule::ARule)::Tuple{Itemset,Itemset}

Getter for the content of an ARule, that is, both its antecedent and its consequent.

See also antecedent, ARule, consequent, Itemset.

SoleRules.antecedent — Method

antecedent(rule::ARule)::Itemset

Getter for rule's antecedent.

See also antecedent, ARule, Itemset.

SoleRules.consequent — Method

consequent(rule::ARule)::Itemset

Getter for rule's consequent.

See also consequent, ARule, Itemset.

When you want to consider a generic entity obtained through an association rule mining algorithm (frequent itemsets and, of course, association rules) just use the following type.

SoleRules.ARMSubject — Type

ARMSubject = Union{ARule,Itemset}

Memoizable types for association rule mining (ARM).

See also GmeasMemo, GmeasMemoKey, LmeasMemo, LmeasMemoKey.

Measures

We cannot establish when a ARMSubject is interesting just by looking at its shape: we need meaningfulness measures.

SoleRules.Threshold — Type

const Threshold = Float64

Threshold value for meaningfulness measures.

See also gconfidence, gsupport, lconfidence, lsupport.

SoleRules.MeaningfulnessMeasure — Type

const MeaningfulnessMeasure = Tuple{Function, Threshold, Threshold}

In the classic propositional case scenario where each instance of a dataset is composed of just a single world (it is a propositional interpretation), a meaningfulness measure is simply a function which measures how many times a property of an Itemset or an ARule is respected across all instances of the dataset.

In the context of modal logic, where the instances of a dataset are relational objects, every meaningfulness measure must capture two aspects: how much an Itemset or an ARule is meaningful inside an instance, and how much the same object is meaningful across all the instances.

For this reason, we can think of a meaningfulness measure as a matryoshka composed of an external global measure and an internal local measure. The global measure tests for how many instances a local measure overpass a local threshold. At the end of the process, a global threshold can be used to establish if the global measure is actually meaningful or not. (Note that this generalizes the propositional logic case scenario, where it is enough to just apply a measure across instances.)

Therefore, a MeaningfulnessMeasure is a tuple composed of a global meaningfulness measure, a local threshold and a global threshold.

See also gconfidence, gsupport, lconfidence, lsupport.

SoleRules.islocalof — Method

islocalof(::Function, ::Function)::Bool

Twin method of isglobalof.

Trait to indicate that a local meaningfulness measure is used as subroutine in a global measure.

For example, islocalof(lsupport, gsupport) is true, and isglobalof(gsupport, lsupport) is false.

Warning

When implementing a custom meaningfulness measure, make sure to implement both traits if necessary. This is fundamental to guarantee the correct behavior of some methods, such as getlocalthreshold.

See also getlocalthreshold, gsupport, isglobalof, lsupport.

SoleRules.isglobalof — Method

isglobalof(::Function, ::Function)::Bool

Twin trait of islocalof.

See also getlocalthreshold, gsupport, islocalof, lsupport.

The following are little data structures which will return useful later, when you will read about how a dataset is "mined", looking for ARMSubjects.

SoleRules.LmeasMemoKey — Type

const LmeasMemoKey = Tuple{Symbol,ARMSubject,Int64}

Key of a LmeasMemo dictionary. Represents a local meaningfulness measure name (as a Symbol), a ARMSubject, and the number of a dataset instance where the measure is applied.

See also LmeasMemo, ARMSubject.

SoleRules.LmeasMemo — Type

const LmeasMemo = Dict{LmeasMemoKey,Threshold}

Association between a local measure of a ARMSubject on a specific dataset instance, and its value.

See also LmeasMemoKey, ARMSubject.

SoleRules.GmeasMemoKey — Type

const GmeasMemoKey = Tuple{Symbol,ARMSubject}

Key of a GmeasMemo dictionary. Represents a global meaningfulness measure name (as a Symbol) and a ARMSubject.

See also GmeasMemo, ARMSubject.

SoleRules.GmeasMemo — Type

const GmeasMemo = Dict{GmeasMemoKey,Threshold}

Association between a global measure of a ARMSubject on a dataset, and its value.

The reference to the dataset is not explicited here, since GmeasMemo is intended to be used as a memoization structure inside Miner objects, and the latter already knows the dataset they are working with.

See also GmeasMemoKey, ARMSubject.

What follows is a list of the already built-in meaningfulness measures. In the hands-on section you will learn how to implement your own measure.

SoleRules.lsupport — Method

function lsupport(
    itemset::Itemset,
    logi_instance::LogicalInstance;
    miner::Union{Nothing,Miner}=nothing
)::Float64

Compute the local support for the given itemset in the given logi_instance.

Local support is the ratio between the number of worlds in a LogicalInstance where and Itemset is true and the total number of worlds in the same instance.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also Miner, LogicalInstance, Itemset.

SoleRules.gsupport — Method

function gsupport(
    itemset::Itemset,
    X::SupportedLogiset,
    threshold::Threshold;
    miner::Union{Nothing,Miner} = nothing
)::Float64

Compute the global support for the given itemset on a logiset X, considering threshold as the threshold for the local support called internally.

Global support is the ratio between the number of LogicalInstances in a SupportedLogiset for which the local support, lsupport, is greater than a Threshold, and the total number of instances in the same logiset.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also Miner, LogicalInstance, Itemset, SupportedLogiset, Threshold.

SoleRules.lconfidence — Method

function lconfidence(
    rule::ARule,
    logi_instance::LogicalInstance;
    miner::Union{Nothing,Miner} = nothing
)::Float64

Compute the local confidence for the given rule in the given logi_instance.

Local confidence is the ratio between lsupport of an ARule on a LogicalInstance and the lsupport of the antecedent of the same rule.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also antecedent, ARule, Miner, LogicalInstance, lsupport.

SoleRules.gconfidence — Method

function gconfidence(
    rule::ARule,
    X::SupportedLogiset,
    threshold::Threshold;
    miner::Union{Nothing,Miner} = nothing
)::Float64

Compute the global confidence for the given rule on a logiset X, considering threshold as the threshold for the global support called internally.

Global confidence is the ratio between gsupport of an ARule on a SupportedLogiset and the gsupport of the antecedent of the same rule.

If a miner is provided, then its internal state is updated and used to leverage memoization.

See also antecedent, ARule, Miner, gsupport, SupportedLogiset.

Mining structures

Finally, we are ready to start mining. To do so, we need to create a Miner object. We just need to specify which dataset we are working with, together with a mining function, a vector of initial Items, and the `MeaningfulnessMeasures to establish ARMSubject interestingness.

SoleRules.Miner — Type

struct Miner{
    D<:AbstractDataset,
    F <:Function,
    I<:Item,
    IM<:MeaningfulnessMeasure,
    RM<:MeaningfulnessMeasure
}
    X::D                            # target dataset
    algorithm::F                    # algorithm used to perform extraction
    items::Vector{I}

                                    # meaningfulness measures
    item_constrained_measures::Vector{IM}
    rule_constrained_measures::Vector{RM}

    freqitems::Vector{Itemset}      # collected frequent itemsets
    arules::Vector{ARule}           # collected association rules

    lmemo::LmeasMemo                # local memoization structure
    gmemo::GmeasMemo                # global memoization structure

    powerups::Powerup               # mining algorithm powerups (see documentation)
    info::Info                      # general informations
end

Machine learning model interface to perform association rules extraction.

Examples

julia> using SoleRules
julia> using SoleData

# Load NATOPS DataFrame
julia> X_df, y = load_arff_dataset("NATOPS");

# Convert NATOPS DataFrame to a Logiset
julia> X = scalarlogiset(X_df)

# Prepare some propositional atoms
julia> p = Atom(ScalarCondition(UnivariateMin(1), >, -0.5))
julia> q = Atom(ScalarCondition(UnivariateMin(2), <=, -2.2))
julia> r = Atom(ScalarCondition(UnivariateMin(3), >, -3.6))

# Prepare modal atoms using later relationship - see [`SoleLogics.IntervalRelation`](@ref))
julia> lp = box(IA_L)(p)
julia> lq = diamond(IA_L)(q)
julia> lr = boxlater(r)

# Compose a vector of items, regrouping the atoms defined before
julia> manual_alphabet = Vector{Item}([p, q, r, lp, lq, lr])

# Create an association rule miner wrapping `fpgrowth` algorithm - see [`fpgrowth`](@ref);
# note that meaningfulness measures are not explicited and, thus, are defaulted as in the
# call below.
julia> miner = Miner(X, fpgrowth(), manual_alphabet)

# Create an association rule miner, expliciting global meaningfulness measures with their
# local and global thresholds, both for [`Itemset`](@ref)s and [`ARule`](@ref).
julia> miner = Miner(X, fpgrowth(), manual_alphabet,
    [(gsupport, 0.1, 0.1)], [(gconfidence, 0.2, 0.2)])

# Consider the dataset and learning algorithm wrapped by `miner` (resp., `X` and `fpgrowth`)
# Mine the frequent itemsets, that is, those for which item measures are large enough.
# Then iterate the generator returned by [`mine`](@ref) to enumerate association rules.
julia> for arule in SoleRules.mine(miner)
    println(miner)
end

See also ARule, apriori, MeaningfulnessMeasure, Itemset, GmeasMemo, LmeasMemo.

SoleRules.dataset — Method

dataset(miner::Miner)::AbstractDataset

Getter for the dataset wrapped by miners.

See SoleBase.AbstractDataset, Miner.

SoleRules.algorithm — Method

algorithm(miner::Miner)::Function

Getter for the mining algorithm loaded into miner.

See Miner.

SoleRules.items — Method

items(miner::Miner)

Getter for the items of Items loaded into miner.

See Item, Miner.

SoleRules.item_meas — Method

item_meas(miner::Miner)::Vector{<:MeaningfulnessMeasure}

Return the MeaningfulnessMeasures tailored to work with Itemsets, loaded inside miner.

See Itemset, MeaningfulnessMeasure, Miner.

SoleRules.rule_meas — Method

rule_meas(miner::Miner)::Vector{<:MeaningfulnessMeasure}

Return the MeaningfulnessMeasures tailored to work with ARules, loaded inside miner.

See Miner, ARule, MeaningfulnessMeasure.

SoleRules.getlocalthreshold — Method

getlocalthreshold(miner::Miner, meas::Function)::Threshold

Getter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work locally (that is, analyzing "the inside" of a dataset's instances) in miner.

See Miner, MeaningfulnessMeasure, Threshold.

SoleRules.setlocalthreshold — Method

setlocalthreshold(miner::Miner, meas::Function, threshold::Threshold)

Setter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work locally (that is, analyzing "the inside" of a dataset's instances) in miner.

See Miner, MeaningfulnessMeasure, Threshold.

SoleRules.getglobalthreshold — Method

getglobalthreshold(miner::Miner, meas::Function)::Threshold

Getter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work globally (that is, measuring the behavior of a specific local-measure across all dataset's instances) in miner.

See Miner, MeaningfulnessMeasure, Threshold.

SoleRules.setglobalthreshold — Method

setglobalthreshold(miner::Miner, meas::Function, threshold::Threshold)

Setter for the Threshold associated with the function wrapped by some MeaningfulnessMeasure tailored to work globally (that is, measuring the behavior of a specific local-measure across all dataset's instances) in miner.

See Miner, MeaningfulnessMeasure, Threshold.

After a Miner ends mining, frequent Itemsets and ARule are accessibles through the getters below.

SoleRules.freqitems — Method

freqitems(miner::Miner)

Return all frequent Itemsets mined by miner.

See also Itemset, Miner.

SoleRules.arules — Method

arules(miner::Miner)

Return all the ARules mined by miner.

See also ARule, Miner.

SoleRules.localmemo — Method

localmemo(miner::Miner)::LmeasMemo
localmemo(miner::Miner, key::LmeasMemoKey)

Return the local memoization structure inside miner, or a specific entry if a LmeasMemoKey is provided.

See also Miner, LmeasMemo, LmeasMemoKey.

SoleRules.localmemo! — Method

localmemo!(miner::Miner, key::LmeasMemoKey, val::Threshold)

Setter for a specific entry key inside the local memoization structure wrapped by miner.

See also Miner, LmeasMemo, LmeasMemoKey.

SoleRules.globalmemo — Method

globalmemo(miner::Miner)::GmeasMemo
globalmemo(miner::Miner, key::GmeasMemoKey)

Return the global memoization structure inside miner, or a specific entry if a GmeasMemoKey is provided.

See also Miner, GmeasMemo, GmeasMemoKey.

SoleRules.globalmemo! — Method

globalmemo!(miner::Miner, key::GmeasMemoKey, val::Threshold)

Setter for a specific entry key inside the global memoization structure wrapped by miner.

See also Miner, GmeasMemo, GmeasMemoKey.

The info field in Miner is a dictionary used to store extra informations about the miner, such as statistics about mining. Currently, since the package is still being developed, the info field only contains a flag indicating whether the miner has been used for mining or no.

SoleRules.info — Method

info(miner::Miner)::Powerup
info(miner::Miner, key::Symbol)

Getter for the entire additional informations field inside a miner, or one of its specific entries.

See also Miner, Powerup.

SoleRules.info! — Method

info!(miner::Miner, key::Symbol, val)

Setter for the content of a specific field of miner's info.

See also hasinfo, info, Miner.

SoleRules.hasinfo — Method

hasinfo(miner::Miner, key::Symbol)

Return whether miner additional informations field contains an entry key.

See also Miner.

When writing your own mining algorithm, or when mining with a particular kind of dataset, you might need to specialize the Miner, keeping, for example, custom meta data and data structures. To specialize a Miner, you can fill a Powerup structure to fit your needs.

SoleRules.Powerup — Type

const Powerup = Dict{Symbol,Any}

Additional informations associated with an ARMSubject that can be used to specialize a Miner, augmenting its capabilities.

To understand how to specialize a Miner, see haspowerup, initpowerups, 'powerups`, powerups!.

SoleRules.powerups — Method

powerups(miner::Miner)::Powerup
powerups(miner::Miner, key::Symbol)

Getter for the entire powerups structure currently loaded in miner, or a specific powerup.

See also haspowerup, initpowerups, Miner, Powerup.

SoleRules.powerups! — Method

powerups!(miner::Miner, key::Symbol, val)

Setter for the content of a specific field of miner's powerups.

See also haspowerup, initpowerups, Miner, Powerup.

SoleRules.haspowerup — Method

haspowerup(miner::Miner, key::Symbol)

Return whether miner powerups field contains an entry key.

See also Miner, Powerup, powerups.

SoleRules.initpowerups — Method

initpowerups(::Function, ::AbstractDataset)

This defines how Miner's powerup field is filled to optimize the mining.

To conclude this section, this is how to start mining.

SoleRules.mine — Method

mine(miner::Miner)

Synonym for SoleRules.apply(miner, dataset(miner)).

See also ARule, Itemset, SoleRules.apply.

SoleRules.apply — Method

apply(miner::Miner, X::AbstractDataset)

Extract association rules in the dataset referenced by miner, saving the interesting Itemsets inside miner. Then, return a generator of ARules.

See also ARule, Itemset.