Getting started

Core definitions

Measures

Mining structures

Miner customization

Utilities

The following utilities often involve performing some combinatoric trick between Itemsets and ARules, and might be useful to avoid reinventing the wheel.

Items and Itemsets

ModalAssociationRules.combine_items — Function

combine_items(itemsets::Vector{<:Itemset}, newlength::Integer)

Return a generator which combines Itemsets from itemsets into new itemsets of length newlength by taking all combinations of two itemsets and joining them.

See also Itemset.

combine_items(variable::Vector{<:Item}, fixed::Vector{<:Item})

Return a generator of Itemset, which iterates the combinations of Items in variable and prepend them to fixed vector.

See also Item, Itemset.

ModalAssociationRules.grow_prune — Function

grow_prune(candidates::Vector{Itemset}, frequents::Vector{Itemset}, k::Integer)

Return a generator, which yields only the candidates for which every (k-1)-length subset is in frequents.

See also Itemset.

Missing docstring.

Missing docstring for coalesce_contributors(itemset::Itemset, miner::Miner; lmeas::Function=lsupport). Check Documenter's build log for details.

Association rules

ModalAssociationRules.arules_generator — Method

arules_generator(itemsets::Vector{Itemset}, miner::Miner)

Raw subroutine of generaterules!(miner::Miner; kwargs...).

Generates ARule from the given collection of itemsets and miner.

The strategy followed is described here at section 2.2.

To establish the meaningfulness of each association rule, check if it meets the global constraints specified in rulemeasures(miner), and yields the rule if so.

See also ARule, Miner, Itemset, rulemeasures.